Rename 'vec2D' to 'vec2'.
